src/requirements.txt
torchdata
diffusers
hydra-core
omegaconf
tqdm
accelerate
wandb
einops
transformers
alias_free_torch
h5py
torchaudio
soundfile
tensorboard#add
botocore
s3transfer
sentencepiece
openai
laion-clap
pysptk
pyworld
sed_eval
dcase_util